# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  description: Descripción de la API
  contact:
    email: tidy@tidy.com
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: http://localhost:5173
    description: Servidor del frontend
  - url: http://tidy-app.me
    description: Ruta de dominio para enviar correos electrónicos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Activity:
      type: object
      properties:
        id:
          type: integer
          description: ID de la actividad
        titulo:
          type: string
          description: Título de la actividad
        description:
          type: string
          description: Descripción de la actividad
        fecha_inicio:
          type: string
          format: date-time
          description: Fecha de inicio de la actividad
        fecha_fin:
          type: string
          format: date-time
          description: Fecha de fin de la actividad
        estado:
          type: string
          enum: [pendiente, en_progreso, completada]
          description: Estado de la actividad
        num_preguntas:
          type: integer
          description: Número de preguntas de la actividad
        prioridad_id:
          type: integer
          description: ID de la prioridad de la actividad
        user_id:
          type: integer
          description: ID del usuario
        subject_id:
          type: integer
          description: ID de la materia
    CreateActivityRequest:
      type: object
      properties:
        titulo:
          type: string
          description: Título de la actividad
        description:
          type: string
          description: Descripción de la actividad
        fecha_inicio:
          type: string
          format: date-time
          description: Fecha de inicio de la actividad
        fecha_fin:
          type: string
          format: date-time
          description: Fecha de fin de la actividad
        estado:
          type: string
          enum: [pendiente, en_progreso, completada]
          description: Estado de la actividad
        num_preguntas:
          type: integer
          description: Número de preguntas de la actividad
        prioridad_id:
          type: integer
          description: ID de la prioridad de la actividad
    Subtask:
      type: object
      properties:
        id:
          type: integer
          description: ID de la subtarea
        titulo:
          type: string
          description: Título de la subtarea
        description:
          type: string
          description: Descripción de la subtarea
        estado:
          type: string
          enum: [pendiente, en_progreso, completada]
          description: Estado de la subtarea
        actividad_id:
          type: integer
          description: ID de la actividad
...
security:
  - bearerAuth: []
paths:
...
  /request-password-reset:
    post:
      summary: Solicitar restablecimiento de contraseña
      description: Permite solicitar un restablecimiento de contraseña.
      operationId: requestPasswordReset
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Correo electrónico del usuario
      responses:
        '200':
          description: Solicitud de restablecimiento enviada
        '400':
          description: Error en la solicitud
  /reset-password/{token}:
    post:
      summary: Restablecer contraseña
      description: Permite restablecer la contraseña de un usuario.
      operationId: resetPassword
      tags:
        - Auth
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  description: Nueva contraseña del usuario
      responses:
        '200':
          description: Contraseña restablecida exitosamente
        '400':
          description: Error en la solicitud
  /verify-reset-token/{token}:
    get:
      summary: Verificar token de restablecimiento de contraseña
      description: Permite verificar la validez de un token de restablecimiento de contraseña.
      operationId: verifyResetToken
      tags:
        - Auth
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token válido
        '400':
          description: Token inválido
  /verify-email/{token}:
    get:
      summary: Verificar correo electrónico
      description: Permite verificar la validez de un token de verificación de correo electrónico.
      operationId: verifyEmail
      tags:
        - Auth
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Correo electrónico verificado exitosamente
        '400':
          description: Token inválido
...
  /subjects/{id}/activities:
    get:
      summary: Obtener todas las actividades de una materia específica
      description: Permite obtener todas las actividades asociadas a una materia específica.
      operationId: getActivities
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de actividades obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '401':
          description: No autorizado

    post:
      summary: Crear una nueva actividad en una materia específica
      description: Permite crear una nueva actividad asociada a una materia específica.
      operationId: createActivity
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Actividad creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado
  /subjects/{id}/activities/{idActivity}:
    get:
      summary: Obtener una actividad específica de una materia
      description: Permite obtener una actividad específica asociada a una materia.
      operationId: getActivityById
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: idActivity
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Actividad obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '401':
          description: No autorizado
        '404':
          description: No se encontró la actividad

    put:
      summary: Actualizar una actividad existente
      description: Permite actualizar una actividad existente asociada a una materia.
      operationId: updateActivity
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: idActivity
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '200':
          description: Actividad actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado

    delete:
      summary: Eliminar una actividad existente
      description: Permite eliminar una actividad existente asociada a una materia.
      operationId: deleteActivity
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: idActivity
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Actividad eliminada exitosamente
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado
...
  /activities/{idActivity}/subtasks:
    get:
      summary: Obtener todas las subtareas de una actividad específica
      description: Permite obtener todas las subtareas asociadas a una actividad específica.
      operationId: getSubtasks
      tags:
        - Subtasks
      security:
        - bearerAuth: []
      parameters:
        - name: idActivity
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de subtareas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subtask'
        '401':
          description: No autorizado
...